added support for date presets

+++ database/migrations/2022_02_23_160900_create_ad_sets_table.php
@@ -13,7 +13,7 @@ return new class extends FacebookMarketingApiMigration {
     public function up()
     {
         Schema::create('facebook_ad_sets', function (Blueprint $table) {
-            $table->integer('id')->primary();
+            $table->bigInteger('id')->primary();
             $table->string('name');
             $table->timestamps();
         });
